Modified DbApi and added deleteAllTasks and deleteAllComment

+++ Database/dbapi.cpp
@@ -3,11 +3,6 @@
 bool DbApi::isOpen = false;
 QSqlDatabase DbApi::db = QSqlDatabase::addDatabase("QSQLITE");
 
-DbApi::DbApi()
-{
-
-}
-
 bool DbApi::openDb(QString pathToDb){
     if(isOpen)
         db.close();
@@ -15,8 +10,10 @@ bool DbApi::openDb(QString pathToDb){
     db.setDatabaseName(pathToDb);
     isOpen = (db.open()) ? true : false;
 
-    if(!isOpen)
+    if(!isOpen){
+        qDebug() << "all is bad";
         return false;
+    }
 
     QString queryStr("CREATE TABLE IF NOT EXISTS \"task\" ("
         "\"id\"	INTEGER NOT NULL UNIQUE,"
@@ -32,7 +29,7 @@ bool DbApi::openDb(QString pathToDb){
 
     queryStr = "CREATE TABLE IF NOT EXISTS \"comment\" ("
         "\"id\"	INTEGER NOT NULL UNIQUE,"
-        "\"id_task\"	INTEGER NOT NULL UNIQUE,"
+        "\"id_task\"	INTEGER NOT NULL,"
         "\"comment\"	TEXT NOT NULL,"
         "FOREIGN KEY(\"id_task\") REFERENCES \"task\"(\"id\") ON DELETE CASCADE,"
         "PRIMARY KEY(\"id\" AUTOINCREMENT)"
@@ -42,7 +39,32 @@ bool DbApi::openDb(QString pathToDb){
     return true;
 }
 
+void DbApi::closeDb(){
+    db.close();
+}
+
+Task* DbApi::getTaskById(unsigned int id){
+    if(!isOpened())
+        return nullptr;
+    QSqlQuery query(nullptr, db);
+    QString queryStr = "SELECT * FROM task WHERE id=%1;";
+    Task *gettedTask = new Task();
+    queryStr = queryStr.arg(id);
+    query.exec(queryStr);
+    QSqlRecord rec = query.record();
+
+    while(query.next()){
+        gettedTask->setTaskId(query.value(rec.indexOf("id")).toUInt());
+        gettedTask->setTaskName(query.value(rec.indexOf("name")).toString().toStdString());
+        //gettedTask->setDescription(query.value(rec.indexOf("description")).toString().toStdString());
+    }
+
+    return gettedTask;
+}
+
 bool DbApi::insertTask(const Task &task){
+    if(!isOpened())
+        return false;
     QSqlQuery query;
     QString queryStr = "INSERT INTO task(name, description, status) "
                 "VALUES ('%1', '%2', '%3');";
@@ -54,6 +76,8 @@ bool DbApi::insertTask(const Task &task){
 }
 
 bool DbApi::updateTask(const Task &task){
+    if(!isOpened())
+        return false;
     QSqlQuery query;
     QString queryStr = "UPDATE task SET name='%1', description='%2', status='%3' WHERE id=%4;";
     queryStr = queryStr.arg(task.getTaskName().c_str());
@@ -65,13 +89,27 @@ bool DbApi::updateTask(const Task &task){
 }
 
 void DbApi::deleteTaskById(unsigned int id){
+    if(!isOpened())
+        return;
     QSqlQuery query;
     QString queryStr = "DELETE FROM task WHERE id=%1;";
     queryStr = queryStr.arg(id);
     query.prepare(queryStr);
+    query.exec();
+}
+
+void DbApi::deleteAllTasks(){
+    if(!isOpened())
+        return;
+    QSqlQuery query;
+    QString queryStr = "DELETE FROM task;";
+    query.prepare(queryStr);
+    query.exec();
 }
 
 bool DbApi::insertComment(const Task &task, std::string comment){
+    if(!isOpened())
+        return false;
     QSqlQuery query;
     QString queryStr = "INSERT INTO comment(id_task, comment) "
                 "VALUES (%1, '%2');";
@@ -82,8 +120,30 @@ bool DbApi::insertComment(const Task &task, std::string comment){
 }
 
 void DbApi::deleteCommentById(unsigned int id){
+    if(!isOpened())
+        return;
     QSqlQuery query;
     QString queryStr = "DELETE FROM comment WHERE id=%1;";
     queryStr = queryStr.arg(id);
     query.prepare(queryStr);
+    query.exec();
+}
+
+void DbApi::deleteAllComments(){
+    if(!isOpened())
+        return;
+    QSqlQuery query;
+    QString queryStr = "DELETE FROM comment;";
+    query.prepare(queryStr);
+    query.exec();
+}
+
+void DbApi::deleteAllCommentsInTask(unsigned int id_task){
+    if(!isOpened())
+        return;
+    QSqlQuery query;
+    QString queryStr = "DELETE FROM comment WHERE id_task=%1;";
+    queryStr = queryStr.arg(id_task);
+    query.prepare(queryStr);
+    query.exec();
 }
